refactor: attempted to fix the Norminette
I have tried to fix the Norm errors in the code as
much as possible. The errors that remain are that
which affect the code and logic of the project.
These would need to be removed with careful
consideration.

Furthermore, I recommend that we stick to writing
code that is in accordance to the Norminette to
prevent us from having unnecessary problems down
the road.