feat: two stores of env vars are now available
The first store is an array of character pointers
that are a heap duplicate of the original
environment variables that are received. This will
be used for env printing as well as a parameter
for the execve function. Since it is malloced, it
will be easier to add and remove to it through a
personal implementation of realloc.

The second store is the previously created linked
list. Though the feature for that is that now I
sort the variables alphabetically before placing
them in the linked list. This will help with the
specific printing behavior of export. Addition and
removal of it will also be simple through the
manipulation of the next pointers in the nodes and
allocating for new ones.