feat: added the unset builtin
I added the unset builtin function. It works by
identifying the variable in question within the
two lists of the environment variables and
deletes/frees the variable and content. For the
changes matrix list, it alters the pointers in a
manner similar to lists, where the deleted entry
is dropped from the indexing.