feat: wildcards now expand to the files in the cwd
For every token of type WORD, I check for a [ * ].
Three system calls are used mainly to identify and
match file names with the token. It ensures that
the hidden files are ignored unless specified.

To achieve the complex pattern matching, I
recursively call the function responsible for it
with updated arguments, so that it essentially
matches the leading characters. If the pattern
needs a comparison of the trailing chararacters,
I use the ft_strncmp() function for that match.

While testing, I came across a bug in the dollar
expansion section. The expansion would ignoring
everything after a [ ' ] regardless of it being
nested within [ " " ] or whether or not there was
a closing [ ' ]. This is now rectified.