fix: hidden files and trailers in wildcards
There was a major issue with the hidden file check
which was due to the hard coding of the cases. Now
that I realised that it couldn't possibly be hard
coded, I have fixed that and concised it into a
shorter form that now fits in the match_leaders().

Now I have removed the old function meant for the
hidden files and replaced it with a function that
instead checks the trailing characters, as that
was the main issue that I needed to tackle. This
meant that I can fix the problems while sticking
to the five functions in wildcard.c