fix: opendir segfault in a non existant directory
Whenever the wildcard would try to expand in a
directory that has been deleted, opendir() calls
segfault due to it trying to read the string that
is still uninitialised. This was fixed by simply
initialising the first index of the string to the
null character and splitting the getcwd() and
opendir() calls onto seperate lines.