refactor: fixed norm errors and added null checks
I mainly formatted the files to be norm friendly.
Any remaining error require new features or a
rearrangement of the code.

The null checks are added to the builtins because
I changed the paramters to accept the same set of
char pointers that exeecve() will accept. This can
result in segfaults which I now protect.