feat: the wildcard adds filenames as tokens
Originally it would join all of the filenames into
a single string and use that as the content for
that specific token. Now I have changed it so that
for each filename that it matches, it will create
a token specifically to store that filename. Once
the token is created, it will be added to the list
of tokens present in the minishell master struct.

Along with that, I made some changes to the order
of the functions to be alphabetically ordered in
terms of the function name.

Comments have been added to all the new functions
that have been created as well.