feat: changed the redirection parsing
A big change made to the parsing is the apparent
removal of everything heredoc. It is no longer
needed as it has been implemented in the
redirection aspect of the parsing. All the opening
of the files for redirection have also been moved
to the parsing side.

The t_cmd struct from the main() function has been
relocated to the minishell struct. The free_cmd()
function will also close all the files that have
been opened.

All the execution with regards to the redirection
has been commented out. This is done to implement
the new aspect of the parsing. Instead of those
functions, we can use the duplicate_fds() function
that I wrote. WARNING: This is currently a WIP.
Allow me to figure out the logic to implement it.

Once I finish this, I will work on implementing
pipes. Currently the changes just reflects the
addition of the pipefds to the minishell struct,
rather than it being local to the function, which
allow for easier fd duplication.

After I get your code for the bonus, I will
assimilate that as well.