fix: path id and command redirection is fixed
The find command function was trying to split the
key [ PATH ] instead of its value, causing failure
in finding commands.

Redirection of the commands was failing due to the
new implementation attempting to read the
non-existant tokens of the redirection struct. It
now confirms first the presence of the execution
struct before it reads its tokens.