feat: Implemented pipes and boolean operators
Added the logic for the pipes and boolean
operators to the execution. The logic is likely
to work without the parenthesis for now. That will
require more development after the parsing errors
are fixed. Organized the files in the execution
sources as well.

A new test case is also added to [ test_cases ].