feat: added the exit builtin
The exit builtin handles the limit of the argument
needing to be the only argument, of it being
numerical as well as remaining within the limits
of long min and long max.

Alongside that, the bugs in cd, unset and export
have also been fixed. That marks all the builtins
to be completed.